Implement a basic bus with end-to-end event dissemination and example
Implement a bus with basic functionality that provides instantiation of
producers and consumers, and disseminates events produced by consumers
to consumers that are interested in them.

Implement basic primitive types from the interface, just enough to allow
implementing an example.

Implement an example that demonstrates bus instantiation with distinct
and overlapping interest. Show that events published by the producer are
received by the consumers. Note, due to concurrency the example output
is unordered.

Leave TODOs as necessary; this implementation is very much a work in
progress.

Depends on #3